crawler/requirements.txt
requests
beautifulsoup4
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
